pyproject.toml
[tool.poetry]
name = "expert-umbrella"
version = "0.1.0"
description = ""
authors = ["alphabeta"]
readme = "README.md"
packages = [{include = "expert_umbrella"}]

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.0"
blue = "^0.9.1"
django-debug-toolbar = "^4.2.0"
django-notifications-hq = "^1.8.3"

[tool.taskipy.tasks]
server = "python manage.py runserver"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
test = "python manage.py test"
collectstatic =  "python manage.py collectstatic"
createsuperuser =  "python manage.py createsuperuser"
loaddata =  "python manage.py loaddata"
createcachetable = "python manage.py createcachetable"
format = "blue -S ."

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.7"
django-widget-tweaks = "^1.5.0"
markdown = "^3.5.1"
django-gravatar2 = "^1.4.4"
bleach = {extras = ["css"], version = "^6.1.0"}
django-select2 = "^8.1.2"
python-dotenv = "^1.0.0"
mysqlclient = "^2.2.0"
whitenoise = "^6.6.0"
django-compressor = "^4.4"
gunicorn = "^21.2.0"
django-simple-captcha = "^0.5.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
